Rename modules to lowercase and update version to 0.2.7 in setup.py
