+++ src/modules/excel_report.py
@@ -15,12 +15,14 @@ Functions:
 """
 
 import os
+import subprocess
+
 from colorama import Fore, Style
 from openpyxl.reader.excel import load_workbook
 import pandas as pd
 from pandas import ExcelWriter
 from tabulate import tabulate
-import src.utils.Settings as Settings
+import src.utils.settings_definitions as Settings
 
 
 class ExcelReport:
@@ -67,7 +69,8 @@ class ExcelReport:
 
 def try_open(final_out_path, data_obj):
     """
-    Tries to open the Excel report at the specified path. If it fails, it prints the report in the terminal.
+    Tries to open the Excel report at the specified path. If it fails,
+     it prints the report in the terminal.
 
     :param final_out_path: The path of the Excel report to be opened.
     :param data_obj: pandas.DataFrame: The data object containing the report data.
@@ -75,8 +78,9 @@ def try_open(final_out_path, data_obj):
     """
     try:
         print("Open Excel report: " + final_out_path)
-        os.startfile(final_out_path)
-    except Exception as e:
+        #os.startfile(final_out_path)
+        subprocess.call(['open', final_out_path])
+    except (OSError, IOError) as e:
         print("Error opening Excel report: " + str(e))
         # Sin importar si se puede abrir o no el Excel creado mostramos el reporte en la terminal
         header_list = data_obj.columns
